+++ singularity-eos/eos/eos_carnahan_starling.hpp
@@ -71,9 +71,9 @@ class CarnahanStarling : public EosBase<CarnahanStarling> {
     return std::max(robust::SMALL(), (sie - _qq) / _Cv);
   }
   PORTABLE_INLINE_FUNCTION void CheckParams() const {
-    PORTABLE_ALWAYS_REQUIRE(_Cv >= 0, "Heat capacity must be positive");
-    PORTABLE_ALWAYS_REQUIRE(_gm1 >= 0, "Gruneisen parameter must be positive");
-    PORTABLE_ALWAYS_REQUIRE(_bb >= 0, "Covolume must be positive");
+    PORTABLE_ALWAYS_REQUIRE(_Cv > 0, "Heat capacity must be strictly positive");
+    PORTABLE_ALWAYS_REQUIRE(_gm1 > 0, "Gruneisen parameter must be positive");
+    PORTABLE_ALWAYS_REQUIRE(_bb >= 0, "Covolume must be strictly non-negative");
     _AZbar.CheckParams();
   }
   template <typename Indexer_t = Real *>
@@ -94,24 +94,51 @@ class CarnahanStarling : public EosBase<CarnahanStarling> {
   PORTABLE_INLINE_FUNCTION Real DensityFromPressureTemperature(
       const Real press, const Real temperature, const Real guess = robust::SMALL(),
       Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
-    Real real_root;
-    auto poly = [=](Real dens) {
-      return _Cv * temperature * _gm1 * dens * ZedFromDensity(dens);
+    static constexpr Real rho_low = 0.;
+    // Need to identify an appropriate density scaling for the tolerances.
+    // Note: cannot use `_rho0` since `_rho0` is computed from this function itself.
+    const Real xtol = 1.0e-9;
+    const Real ytol = 1.0e-9;
+
+    // At absolute zero, P = 0, so assumed vacuum state (no cold curve exists).
+    if (temperature <= std::numeric_limits<Real>::min()) {
+      return rho_low;
+    }
+
+    // Use ideal gas value when _bb is zero
+    if (_bb <= std::numeric_limits<Real>::min()) {
+      return robust::ratio(press, _Cv * _gm1 * temperature);
+    }
+    const Real rho_high = 1.0 / _bb;
+
+    // Setup lambda function for finding rho.
+    // The equation has been rewritten to avoid division by zero (polynomial equation).
+    auto f = [=](const Real x /* density */) {
+      const Real eta = _bb * x;
+      const Real term1 = x * (1. + eta + eta * eta - eta * eta * eta);
+      const Real term2 =
+          robust::ratio(press * math_utils::pow<3>(1. - eta), _Cv * _gm1 * temperature);
+      return term1 - term2;
     };
-    using RootFinding1D::findRoot;
-    using RootFinding1D::Status;
-    static constexpr Real xtol = 1.0e-12;
-    static constexpr Real ytol = 1.0e-12;
-    static constexpr Real lo_bound = robust::SMALL();
-    const Real hi_bound = robust::ratio(1.0, _bb);
-    auto status = findRoot(poly, press, guess, lo_bound, hi_bound, xtol, ytol, real_root);
-    if (status != Status::SUCCESS) {
+
+    const RootFinding1D::RootCounts root_info;
+    Real rho; // `rho` will be the root of the equation $f=0$.
+
+    /* The regula falsi method should always return a unique real root in the interval (0,
+     * b^{-1}) since f(0) < 0 and f(b^{-1}) > 0. It can be shown that the derivative of f
+     * (as a function of \eta) is always positive since \eta \in (0,1) hence the root is
+     * unique. */
+    auto status =
+        regula_falsi(f, 0.0 /*target*/, guess, rho_low /*left bracket*/,
+                     rho_high /*right bracket*/, xtol, ytol, rho, &root_info, true);
+
+    if (status != RootFinding1D::Status::SUCCESS) {
       // Root finder failed even though the solution was bracketed... this is an error
       EOS_ERROR("*** (Warning) DensityFromPressureTemperature :: Convergence not met in "
-                "Carnahan-Starling EoS (root finder util) ***\n");
-      real_root = -1.0;
+                "Carnahan-Starling EoS (root finder `regula_falsi`) ***\n");
+      rho = -1.0; // guarantees zero density is returned on convergence failure.
     }
-    return std::max(robust::SMALL(), real_root);
+    return std::max(rho_low, rho);
   }
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real InternalEnergyFromDensityTemperature(